Moved GameMode into pumpkin-core
