+++ pumpkin-core/src/lib.rs
@@ -1,2 +1,5 @@
+pub mod gamemode;
 pub mod random;
 pub mod text;
+
+pub use gamemode::GameMode;
