+++ pumpkin/src/client/player_packet.rs
@@ -3,12 +3,12 @@ use std::f32::consts::PI;
 use crate::{
     commands::{handle_command, CommandSender},
     config::ADVANCED_CONFIG,
-    entity::player::{ChatMode, GameMode, Hand, Player},
+    entity::player::{ChatMode, Hand, Player},
     server::Server,
     util::math::wrap_degrees,
 };
 use num_traits::FromPrimitive;
-use pumpkin_core::text::TextComponent;
+use pumpkin_core::{text::TextComponent, GameMode};
 use pumpkin_entity::EntityId;
 use pumpkin_inventory::WindowType;
 use pumpkin_protocol::server::play::{SCloseContainer, SSetPlayerGround, SUseItem};
